% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key_texture.R
\name{draw_key_texture}
\alias{draw_key_texture}
\title{Key drawing function for textured rectangles}
\usage{
draw_key_texture(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
Key drawing function for textured rectangles
}
