% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtextures.R
\docType{package}
\name{ggtextures}
\alias{ggtextures}
\alias{ggtextures-package}
\title{Tools for drawing textured rectangles and bars with ggplot}
\description{
Tools for drawing textured rectangles and bars with ggplot
}
