% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-image.R
\name{scale_image_identity}
\alias{scale_image_identity}
\alias{scale_image_discrete}
\alias{scale_image_manual}
\title{Image scales}
\usage{
scale_image_identity(..., guide = "none", aesthetics = "image")

scale_image_manual(..., values, aesthetics = "image")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:discrete_scale]{discrete_scale()}} or
\code{\link[ggplot2:continuous_scale]{continuous_scale()}}}

\item{guide}{Guide to use for this scale. Defaults to \code{"none"}.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}
}
\description{
Image scales
}
